Backend Services Unit Tests OK.

+++ src/main/java/org/fpcm/util/passwordmeter/PasswordCheckControls.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class PasswordCheckControls {
 
-	private int nRepInc;
+	private double nRepInc;
 	private int nSeqChar;
 	private int nReqChar;
 }
